chore: Refactor User model to users app
